Added minting for contract transaction

+++ Assets/Scripts/Utils/Balance.cs
@@ -1,5 +1,5 @@
 public class Balance : ResponsePayload
 {
     public decimal balance;
-    public long nonce;
+    public int nonce;
 }
